Fix index.ts and mod.ts mixup. Shims for index.ts on all packages.
